fix(ci): satisfy clippy after memory module move
