+++ src/memory/promote/summary.rs
